Applied sessions and allowed for account login
Got the login page to access the database and log in users based on the
information in the database. Then once the user log in it starts a
session and stores their personal information to be used in the site. I
also changed in header.php to change the login anchor to a logout of
"username" link which links to logout.php which unsets all the session
variables and closes the session