Extended Filtering, Product Details Work
Put the shop sidebar in its own document and made it an include for
easier editing on it; sidebar now works on all pages to direct to the
shop page with correct filtering. Price Range still does not work.
Changed "Shop" dropdown to a single link to shop.php. Added "View
Details" button to link to product details page. THis button also sends
the productID via php to be picked up by product-details.php.
product-details.php now displays the correct item and information
depending on what product was selected